+++ .gitignore
@@ -36,6 +36,3 @@ yarn-error.log*
 next-env.d.ts
 
 .vercel
-
-# keyFile
-src/app/cron/*.json
