+++ next.config.mjs
@@ -8,8 +8,6 @@ const nextConfig = {
     images : {
         domains : [
             'shopping-phinf.pstatic.net',
-            'drive.google.com',
-            'lh3.googleusercontent.com'
         ]
     }
 };
