+++ package.json
@@ -9,22 +9,16 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "google-auth-library": "^9.15.1",
-    "google-spreadsheet": "^4.1.4",
-    "googleapis": "^148.0.0",
-    "next": "15.0.0-rc.0",
-    "node-cron": "^3.0.3",
     "react": "19.0.0-rc-f994737d14-20240522",
     "react-dom": "19.0.0-rc-f994737d14-20240522",
-    "sqlite3": "^5.1.7"
+    "next": "15.0.0-rc.0"
   },
   "devDependencies": {
+    "typescript": "^5",
     "@types/node": "^20",
-    "@types/node-cron": "^3.0.11",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "eslint": "^8",
-    "eslint-config-next": "15.0.0-rc.0",
-    "typescript": "^5"
+    "eslint-config-next": "15.0.0-rc.0"
   }
 }
