--- src/app/cgcProduct/[id]/page.module.css
@@ -1,60 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  gap: 50px;
-}
-
-.container > section {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.cover_img_container {
-  display: flex;
-  justify-content: center;
-  padding: 20px;
-
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
-
-  position: relative;
-}
-
-.cover_img_container::before {
-  position: absolute;
-  top: 0px;
-  left: 0p;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.7);
-  content: "";
-}
-
-.cover_img_container > img {
-  z-index: 1;
-  max-height: 350px;
-  height: 100%;
-}
-
-.title {
-  font-size: large;
-  font-weight: bold;
-}
-
-.subTitle {
-  color: gray;
-}
-
-.author {
-  color: gray;
-}
-
-.description {
-  background-color: rgb(245, 245, 245);
-  padding: 15px;
-  line-height: 1.3;
-  white-space: pre-line;
-  border-radius: 5px;
-}
