+++ src/components/book-item.module.css
@@ -1,4 +1,4 @@
-a.container {
+.container {
   display: flex;
   gap: 15px;
   padding: 20px 10px;
@@ -8,14 +8,8 @@ a.container {
   text-decoration: none;
 }
 
-div.container {
-  display: flex;
-  gap: 15px;
-}
-
 .container img {
-  width: 100px;
-  height: 120px;
+  width: 80px;
 }
 
 .title {
