+++ src/components/searchbar.tsx
@@ -20,11 +20,8 @@ export default function Searchbar() {
   };
 
   const onSubmit = () => {
-    if (!search || q === search) {
-      router.push(`/`);
-      return;
-    } 
-    router.push(`/search?q=${search}&page=1`);
+    if (!search || q === search) return;
+    router.push(`/search?q=${search}`);
   };
 
   const onKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
