+++ src/types.ts
@@ -15,20 +15,3 @@ export interface ReviewData {
   createdAt : string;
   bookId : number;
 }
-
-// --------------------------------------------------------
-
-export interface CGCproduct {
-  id: number;
-  name: string;
-  expirationDate: string;
-  expirationDateNewLot: string;
-  suggestedRetailPrice: string;
-  suggestedWholesalePrice: string;
-  specification: string;
-  category: string;
-  type: string;
-  itemFeatures: string;
-  imageA: string;
-  imageB: string;
-}
